implement decimal fix for issue #4

+++ Models/BaccaratResult.cs
@@ -3,7 +3,7 @@ namespace TeleCasino.BaccaratGameService.Models;
 public class BaccaratResult
 {
     public required string Id { get; set; }
-    public int Wager { get; set; }
+    public decimal Wager { get; set; }
     public decimal Payout { get; set; }
     public decimal NetGain { get; set; }
     public required string VideoFile { get; set; }
@@ -13,8 +13,8 @@ public class BaccaratResult
     public BaccaratBetType BetType { get; set; }
     public List<string> PlayerCards { get; set; } = new();
     public List<string> BankerCards { get; set; } = new();
-    public int PlayerTotal { get; set; }
-    public int BankerTotal { get; set; }
+    public decimal PlayerTotal { get; set; }
+    public decimal BankerTotal { get; set; }
     public BaccaratBetType Winner { get; set; }
     public int GameSessionId { get; set; }
 }
